+++ public/components/all/package.go
@@ -23,7 +23,6 @@ import (
 	_ "github.com/benthosdev/benthos/v4/internal/impl/nats"
 	_ "github.com/benthosdev/benthos/v4/internal/impl/parquet"
 	_ "github.com/benthosdev/benthos/v4/internal/impl/prometheus"
-	_ "github.com/benthosdev/benthos/v4/internal/impl/pulsar"
 	_ "github.com/benthosdev/benthos/v4/internal/impl/redis"
 	_ "github.com/benthosdev/benthos/v4/internal/impl/sql"
 	_ "github.com/benthosdev/benthos/v4/internal/impl/statsd"