+++ public/components/pulsar/package.go
@@ -1,0 +1,3 @@
+package pulsar
+
+import _ "github.com/benthosdev/benthos/v4/internal/impl/pulsar"